+++ bootstrap/ssr/assets/MobileLayout-BQoCFD3J.js
@@ -1,0 +1,9 @@
+import { jsx } from "react/jsx-runtime";
+import "react";
+import { A as Authenticated } from "./AuthenticatedLayout-BpUiURcU.js";
+const MobileLayout = ({ header, children, loading = false }) => {
+  return /* @__PURE__ */ jsx(Authenticated, { header, loading, children });
+};
+export {
+  MobileLayout as M
+};